% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bobyqa_rosen_x1}
\alias{bobyqa_rosen_x1}
\title{Example 1a: Minimize Rosenbrock function using bobyqa}
\usage{
bobyqa_rosen_x1()
}
\value{
No return value, called for side effects.
}
\description{
Minimize Rosenbrock function using bobyqa and
            expect a normal exit from bobyqa.
}
\examples{
fr <- function(x) {   ## Rosenbrock Banana function
 x1 <- x[1]
 x2 <- x[2]
 100 * (x2 - x1 * x1)^2 + (1 - x1)^2
}
(x1 <- minqa::bobyqa(c(1, 2), fr, lower = c(0, 0), upper = c(4, 4)))
## => optimum at c(1, 1) with fval = 0
str(x1) # see that the error code and msg are returned

## corresponding C++ implementation:
bobyqa_rosen_x1()
}
