% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rminqa}
\alias{rminqa}
\title{rminqa}
\description{
Perform derivative-free optimization algorithms in R using C++.
A wrapper interface is provided to call C function of the bobyqa implementation.
}
\author{
Yi Pan, Samuel Watson
}
